[cxx-interop] support sealed c++ interop
This change adds a hidden frontend option that
attempts a best-effort "seal" of use of C++ interop in
a Swift module. This allows clients to build one Swift module
that uses libc++ and C++ interop, and then build another
Swift module that uses the system's c++ standard library and imports
the first module, without errors.

This is done by:
- validating that public APIs don't have C++ types in them.
- validating that inlinable / alwaysEmitIntoClient bodies aren't reliant
  on C++ types.
- erroring out explicitly on type deserialization errors for such
  module, as that allows such module to import C++ Clang modules
  with @implementationOnly, to prevent client Swift modules from
  depending on such C++ modules.
- Setting a new clang macro, that lets imported C++ clang modules
  validate that they're imported in a sealed mode.