[android] build libxml2 and curl with -fPIC
Sets `CMAKE_POSITION_INDEPENDENT_CODE=YES` when building libxml2 and
curl. This causes `-fPIC` to be passed to clang when compiling, which is
required for the i686 Android build to succeed.

This change makes `build.ps1` consistent with the Python build files
swift_build_support\products\libxml2.py and
swift_build_support\products\curl.py, which already define this cmake
property. It is already propertly set for zlib in both locations.