ASTMangler: skip mangling Copyable dependents
Typically, a conformance that is dependent on a conformance to a marker
protocol never reaches this point in the compiler, where we're mangling
the metadata for an opaque return type.

But with the invertible protocols like Copyable, we do permit them, so
we should avoid mangling Copyable as that's generally ABI incompatible
with existing code.

resolves rdar://135310019

(cherry picked from commit 6f6a46f225d29459f66e3e052c2ca33183cdb650)