Serialization: Indicate application extension mode via XCC (#76285)
This fixes the debugger's ability to perform expression evaluation when debugging an 
executable that was built with both under the following circumstances:

1. explicit module build
2. `-application-extension`

The fix is to include `-fapplication-extension` as an XCC field in the swiftmodule. 
This primes the debugger's ClangImporter with the correct flag needed to load the 
explicitly built pcm files generated at build time.