[Backtracer][macOS] Fix crashing thread index when Rosetta is in use.
If you're using Rosetta 2, there are ARM64 threads in your process as
well, which the backtracer skips.  Unfortunately doing that messed up
the crashing thread index, so the backtracer might have crashed or
told you that a different thread was crashing than the one you
thought.

This only affects running x86-64 code on an Apple Silicon Mac.

rdar://135787913