[test] Fix test/IRGen/metadata.swift to match the correct symbol
When this test was updated for new platforms it was accidentally changed
to match the _METACLASS_DATA symbol instead of the _DATA symbol. Fix
that and update the flags to match. This lets us remove the instance
start value 40 from the checks, so the test now matches the comment
saying it should be 8 or 16.

rdar://135453916
(cherry picked from commit cda65d13047a2fb3a45be8327cf977bdd5ee4cfb)