AST: Fix invariant violation in swift::validateGenericSignature()
This is a silly "extended" verification check that we only exercise
in test/Generics/validate_stdlib_generic_signatures.swift.

We were taking a requirement containing unbound dependent member types
and applying a substitution map to it, which would result in an
ErrorType, so the requirement was always considered unsatisfied here
and the check was not as useful as it should have been.

Instead, re-implement a version of isRequirementSatisfied() that only
uses generic signature queries instead of substitution here.